devicehealth: fix the mark out and warn thresholds
The diskprediction plugins all coarsely grade a drive's health
as Good, Warning, or Bad, which we interpret as life expectancy
of >6 weeks, from 2 to 6 weeks, or from 0 to 2 weeks, respectively.

This module does two things: it warns if a drive has a Warning, or
marks an OSD out if a drive is Bad.

So the mark_out_threshold and warn_threshold should be 2 weeks and
6 weeks, respectively.

In 8d7bf7758627ee3f1145e13db761c09be99325c9 the mark_out_threshold
was accidentally doubled from 2 weeks to 4 weeks. Set it back to 2.
Also drop warn_threshold from 12 down to 6 weeks.

Signed-off-by: Dan van der Ster <dan.vanderster@clyso.com>