diskprediction_local: implement a simple rule-based predictor
Add a new disk predictor with a simple rule-based heuristic. If
SMART status is FAILING, the device is Bad. If any well-known
attrs are non-zero, the device gets a Warning. For NVMe devices
we look at the critical_warnings, available_spares, and
percentage_used metrics. Also any device more than 7 years old
will get a Warning.

We will surely tweak these thresolds in the future, but at least
this is comprehensible.

A follow up commit will split this into a separate module so we
can remove the dependency on scipy/sklearn/numpy.

Signed-off-by: Dan van der Ster <dan.vanderster@clyso.com>